refactor: Extract TitledScreen
